+++ test/dummy_tests/application_controller_test.rb
@@ -7,7 +7,7 @@ module ApplicationControllerTests
 
     it 'works' do
       get :index
-      response.body.must_equal '<h1>Rendered MiniTest::Spec</h1>'
+      expect(response.body).must_equal '<h1>Rendered MiniTest::Spec</h1>'
     end
 
     it 'allows custom assertions' do
@@ -15,18 +15,18 @@ module ApplicationControllerTests
     end
 
     it 'can find the controller_class' do
-      self.class.controller_class.must_equal ApplicationController
+      expect(self.class.controller_class).must_equal ApplicationController
     end
 
     it 'can access the setup ivars' do
-      @controller.must_be_kind_of ApplicationController
+      expect(@controller).must_be_kind_of ApplicationController
     end
 
     describe 'nested 1' do
       it('works') { skip }
 
       it 'can find the controller_class' do
-        self.class.controller_class.must_equal ApplicationController
+        expect(self.class.controller_class).must_equal ApplicationController
       end
 
       describe 'nested 2' do
@@ -39,18 +39,18 @@ end
 class ApplicationControllerTest < ActionController::TestCase
   include ApplicationControllerTests
   it 'reflects' do
-    described_class.must_equal ApplicationController
-    self.class.described_class.must_equal ApplicationController
+    expect(described_class).must_equal ApplicationController
+    expect(self.class.described_class).must_equal ApplicationController
   end
   describe 'level 1' do
     it 'reflects' do
-      described_class.must_equal ApplicationController
-      self.class.described_class.must_equal ApplicationController
+      expect(described_class).must_equal ApplicationController
+      expect(self.class.described_class).must_equal ApplicationController
     end
     describe 'level 2' do
       it 'reflects' do
-        described_class.must_equal ApplicationController
-        self.class.described_class.must_equal ApplicationController
+        expect(described_class).must_equal ApplicationController
+        expect(self.class.described_class).must_equal ApplicationController
       end
     end
   end
@@ -59,22 +59,22 @@ end
 describe ApplicationController do
   include ApplicationControllerTests
   it 'class reflects' do
-    described_class.must_equal ApplicationController
-    self.class.described_class.must_equal ApplicationController
+    expect(described_class).must_equal ApplicationController
+    expect(self.class.described_class).must_equal ApplicationController
   end
   it 'reflects' do
-    described_class.must_equal ApplicationController
-    self.class.described_class.must_equal ApplicationController
+    expect(described_class).must_equal ApplicationController
+    expect(self.class.described_class).must_equal ApplicationController
   end
   describe 'level 1' do
     it 'reflects' do
-      described_class.must_equal ApplicationController
-      self.class.described_class.must_equal ApplicationController
+      expect(described_class).must_equal ApplicationController
+      expect(self.class.described_class).must_equal ApplicationController
     end
     describe 'level 2' do
       it 'reflects' do
-        described_class.must_equal ApplicationController
-        self.class.described_class.must_equal ApplicationController
+        expect(described_class).must_equal ApplicationController
+        expect(self.class.described_class).must_equal ApplicationController
       end
     end
   end