+++ test/dummy_tests/library_test.rb
@@ -3,13 +3,13 @@ require "#{Dummy::Application.root}/lib/library"
 
 class LibraryTest < ActiveSupport::TestCase
   it 'reflects' do
-    described_class.must_equal Library
+    expect(described_class).must_equal Library
   end
 end
 
 describe Library do
   it 'reflects' do
-    described_class.must_equal Library
-    self.class.described_class.must_equal Library
+    expect(described_class).must_equal Library
+    expect(self.class.described_class).must_equal Library
   end
 end