+++ test/support/shared_test_case_behavior.rb
@@ -11,10 +11,6 @@ module MiniTestSpecRails
 
     private
 
-    def rails5?
-      Rails.version >= '5.0'
-    end
-
     def setup_dummy_schema
       ActiveRecord::Base.class_eval do
         connection.instance_eval do